% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11openmp-package.R
\docType{package}
\name{cpp11openmp-package}
\alias{cpp11openmp}
\alias{cpp11openmp-package}
\title{cpp11openmp: Using OpenMP with cpp11}
\description{
Provides an example of parallelizing a C++ routine with 'OpenMP' inside a function exposed to R with 'cpp11'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rd-rcpp/cpp11-openmp}
  \item \url{https://rd-rcpp.thecoatlessprofessor.com/cpp11-openmp/}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/cpp11-openmp/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu}
}

}
\keyword{internal}
